+++ ingest/src/main/java/com/microsoft/azure/kusto/ingest/QueuedIngestClient.java
@@ -13,5 +13,5 @@ public interface QueuedIngestClient extends IngestClient {
      * @param queueRequestOptions - Options to use when creating QueueClient
      */
     void setQueueRequestOptions(RequestRetryOptions queueRequestOptions);
-    ResourceManager getResourceManager();
+    ResourceHelper getResourceManager();
 }
